Update site-search functionality
Removes sitesearch.php in favour of a new template function.
Preserves ability to overload with template file if preferred.
Adds function $args to build initially classes for parent in addition to defaults.
Adds new styling for searchform to add a 'wide' class to make form 100% of parent width using flex-box.