Automate BP theme package registration
Commit attempts to hack around the inability to register a theme package from the theme level by adding the BP register function to a file running under 'mu-plugins' (must use).

Creates a new file with bp_register_theme_packages() hook to handle registering progenitors package details as a MU plugin file.

Creates a new function to check for 'mu-plugins' and create the dir if not exists, instantiates the WP filesystem api to then write a new file that simply runs an include for the register file located in '/theme/inc/'

Function hooked on 'admin_init' to run on any action in admin i.e activating the theme or BP having first checked that the file does not already exist.