JS menu scripting update
Add `document.documentElement.clientWidth` to the viewport width check.

Add a removeClass for the toggled class if viewport > 750px - an edge case but if mobile menu open and browser re-sized mobile-menu styling still takes precedence over the normal vert styles as the toggled class isn't removed, this doesn't happen in the real world but safer to remove the toggled class via the resize eventListener and remove the issue.