Update JS menu scripting
Commit adds new approach to re-structering user/group object nav positioning.

Wraps object nav call in the home.php template with a new element shown only if the object nav user selection is set as vert sitea header display.

Scripting adds & removes classes and display property for element to hide or show if we're large of mediung/small screen display.

This ensures that at small to medium screens we display the menu in the item body not at top of screen.