Update members default front page
Commit adds a help/info message when a users WP author description is empty and progenitor is set to show them, hides message leaving edit link when description available.

Creates a new template tag to use to check author meta is 'get'able this replaces the use of get_author_meta checks directly in the templates.