Re-factor object nav item link counts
On object navs in vert main header rendering it's hard to line up text links set to the right when span counts are displayed as they throw the text out of line.

Add new class to the parent list class to state whether there is a span count for list link.

Add spans and classes to text and item count to allow specific styling.

Update styles to reflect the class token names, add margins to text link, modify to half value if item count displayed to force vert line up of text.